requirements.txt
sounddevice==0.4.6
scipy>=1.11.0
openai-whisper>=20231117
pyannote.audio==3.1.1
anthropic>=0.20.0
notion-client==2.2.1
python-dotenv==1.0.0
numpy>=1.24.0
pdfplumber>=0.10.0
python-docx>=1.1.0
openpyxl>=3.1.0
